(edit) create shared/DataTable/utils directory and move the ExpandTableButton, GlobalFilter, PageSizes, SortBy(formaly SortByIcon) components into it
