+++ src/components/shared/DataTable/utils/GlobalFilter.js
